Hijacking the beginning of the engine for physics testing.
Hopefully this will give me an idea of how physics will work in the
context of our engine. Also, it will give me a chance to implement a
cleaner setup of the physics environment.

Since this is my first time working with Box2D I expect to have a few such
iterations. This is the beginning of v0.2.